add prompt optimization example (#467)
* add prompt optimization example


* moved images to a separate repo


* remove image external check


* modify image check


* remove image external check


---------

+++ scripts/check_images.sh
@@ -16,15 +16,9 @@ find content -name "*.md" | while read -r file; do
         img_path=$(echo "$img_tag" | sed -E 's/!\[[^]]*\]\(([^)]+)\)/\1/')
         # Remove any title or size attributes that might be in quotes
         img_path=$(echo "$img_path" | sed -E 's/"[^"]*"//g' | sed -E "s/'[^']*'//g" | xargs)
-        
+
         if [[ "$img_path" == /* ]]; then
             echo "ERROR: $img_path is asbolute path in '$file'"
-            echo "1" > "$temp_error_file"
-            continue
-        fi
-
-        if [[ "$img_path" == https://* ]]; then
-            echo "ERROR: $img_path is external URL in '$file'"
             echo "1" > "$temp_error_file"
             continue
         fi
@@ -32,7 +26,7 @@ find content -name "*.md" | while read -r file; do
         cd "$(dirname "$file")"
 
         # echo "Checking image: $img_path"
-        if [[ ! -f "$img_path" ]]; then
+        if [[ "$img_path" != https://* && -f "$img_path" ]]; then
             echo "1" > "$temp_error_file"
             echo "ERROR: '$img_path' not found in '$file'"
             continue
